==> 02_huvadb_and_testing.Rmd <==
---
title: " 02 - Preparation of the *huva.db* and testing of the dataset"
authors: "Lorenzo Bonaguro"
output: html_document
date: "2022-10-14"
---

# Libraries

## Loading
```{r}
library(huva)
library(huva.db)
```

# Data load

## Loading Data
```{r}
GTEx_dataset <- readRDS("./data/221016_GTEx_dataset_LB.rds")

gene_to_symbol <- readRDS("./data/221016_gene_to_symbol.rds")
```

## Formatting the data
```{r}
gtex_data <- list()

gtex_data[["whole_blood"]] <- as.matrix(GTEx_dataset$`Whole Blood`$data_orig)

gtex_anno <- list()

gtex_anno[["whole_blood"]] <- GTEx_dataset$`Whole Blood`$anno

# Removing columns containing NAs
gtex_anno$whole_blood <- gtex_anno$whole_blood[, c(1,2, 65, 66, 67)]

rownames(gtex_anno$whole_blood) <- gtex_anno$whole_blood$SAMPID

identical(colnames(gtex_data$whole_blood), rownames(gtex_anno$whole_blood))
```

## Inspect the structure
```{r}
str(gtex_data)
```

```{r}
str(gtex_anno)
```

## Prepare the huva.db
```{r}
gtex.db <- generate_huva_dataset(dataset_name = "GTEx_v8", 
                                 data = "gtex_data", 
                                 annotation = "gtex_anno", combine = FALSE)

rm(gtex_anno, gtex_data)
```

```{r}
str(gtex.db)

class(gtex.db)
```

# Session Info
```{r}
info <- sessionInfo()

info
```
